Update Tag testcases
