+++ src/test/java/seedu/fast/model/tag/PriorityTagTest.java
@@ -8,31 +8,8 @@ public class PriorityTagTest {
 
     @Test
     public void constructor_null_throwsNullPointerException() {
-        assertThrows(NullPointerException.class, () -> new Tag(null));
-    }
-
-    @Test
-    public void constructor_invalidTagName_throwsIllegalArgumentException() {
-        String invalidTagName = "";
-        assertThrows(IllegalArgumentException.class, () -> new Tag(invalidTagName));
-    }
-
-    @Test
-    public void isValidTagName() {
-        // null tag name
-        assertThrows(NullPointerException.class, () -> Tag.isValidTagName(null));
-    }
-
-    @Test
-    public void isValidTagTerm() {
-        //null tag term
-        assertThrows(NullPointerException.class, () -> Tag.isValidTagTerm(null));
-    }
-
-    @Test
-    public void createTag() {
-        //null term
-        assertThrows(NullPointerException.class, () -> Tag.createTag(null));
+        assertThrows(NullPointerException.class, () -> new PriorityTag(null));
     }
 
 }
+
